tidy testutils
